added sync enable/disable button to fade editor layout. Move Sync Split Button next to split value label in belt parameter layout. Fixed enable / disable sync setting bugs by using usecases from set-setting in SettingsActions.
